Meta: Move tests modules into `source` folder
