+++ source/github-helpers/search-query.test.ts
@@ -1,7 +1,7 @@
 import test from 'ava';
 
-import './fixtures/globals';
-import SearchQuery from '../source/github-helpers/search-query';
+import '../../fixtures/globals';
+import SearchQuery from './search-query';
 
 test('.get', t => {
 	const query = SearchQuery.from({q: 'wow'});
