+++ source/helpers/loose-parse-int.test.ts
@@ -1,0 +1,9 @@
+import test from 'ava';
+
+import looseParseInt from './loose-parse-int';
+
+test('looseParseInt', t => {
+	t.is(looseParseInt('1,234'), 1234);
+	t.is(looseParseInt('Bugs 1,234'), 1234);
+	t.is(looseParseInt('5000+ issues'), 5000);
+});
