+++ source/helpers/pluralize.test.ts
@@ -1,0 +1,11 @@
+import test from 'ava';
+
+import pluralize from './pluralize';
+
+test('pluralize', t => {
+	t.is(pluralize(0, 'A number', '$$ numbers'), '0 numbers');
+	t.is(pluralize(0, 'A number', '$$ numbers', 'No numbers'), 'No numbers');
+	t.is(pluralize(1, 'A number', '$$ numbers', 'No numbers'), 'A number');
+	t.is(pluralize(2, 'A number', '$$ numbers', 'No numbers'), '2 numbers');
+	t.is(pluralize(2, 'A number', 'Many numbers', 'No numbers'), 'Many numbers');
+});
